refactor: ImageBox 컴포넌트 분리
