+++ frontend/src/components/host/SpaceInfo/index.tsx
@@ -1,7 +1,7 @@
 import useSpaceInfo from './useSpaceInfo';
-import { HiPlus } from 'react-icons/hi';
 
 import Button from '@/components/common/Button';
+import ImageBox from '@/components/host/ImageBox';
 
 import styles from './styles';
 
@@ -12,8 +12,7 @@ interface SpaceInfoProps {
 }
 
 const SpaceInfo: React.FC<SpaceInfoProps> = ({ type = 'read', inputText = '', data }) => {
-  const { imageUrl, labelRef, isActiveSubmit, isExistImageUrl, onChangeImg, onChangeSpaceName, onClickEditSpaceInfo } =
-    useSpaceInfo(data, type);
+  const { isActiveSubmit, onChangeSpaceName, onClickEditSpaceInfo } = useSpaceInfo(data, type);
 
   return (
     <div css={styles.spaceInfo}>
@@ -36,28 +35,7 @@ const SpaceInfo: React.FC<SpaceInfoProps> = ({ type = 'read', inputText = '', da
       <div css={styles.imageContainer}>
         <p css={styles.subTitle}>대표이미지</p>
         <div css={styles.imageWrapper}>
-          {type === 'read' ? (
-            <label css={styles.imageBox(data?.imageUrl)} htmlFor="file" ref={labelRef}></label>
-          ) : (
-            <label css={styles.imageBox(imageUrl, isExistImageUrl ? 'none' : 'dashed')} htmlFor="file" ref={labelRef}>
-              <input
-                css={styles.imageInput}
-                name="imageInput"
-                type="file"
-                id="file"
-                accept="image/*"
-                onChange={onChangeImg}
-              />
-              {!isExistImageUrl && (
-                <div css={styles.iconBox}>
-                  <HiPlus size={50} />
-                </div>
-              )}
-              <p css={styles.imageCoverText}>
-                {isExistImageUrl ? '이미지 수정 시 클릭해 주세요.' : '이미지를 추가해 주세요.'}
-              </p>
-            </label>
-          )}
+          <ImageBox type={type} imageUrl={data?.imageUrl} />
         </div>
       </div>
       <div css={styles.inputContainer}>
