+++ frontend/src/components/host/SpaceInfo/styles.ts
@@ -56,47 +56,6 @@ const imageWrapper = css`
   padding-bottom: 1.25rem;
 `;
 
-const imageBox = (imageUrl: string | undefined, borderStyle?: string) => css`
-  display: block;
-  background: ${theme.colors.gray300};
-  background-size: cover;
-  background-repeat: no-repeat;
-  border: 2px ${theme.colors.gray400};
-  border-radius: 8px;
-  border-style: ${borderStyle};
-  width: 15rem;
-  height: 15rem;
-  position: relative;
-  margin: 0 1.5rem;
-  background-image: url(${imageUrl});
-  cursor: pointer;
-`;
-
-const imageInput = css`
-  opacity: 0;
-  z-index: -1;
-`;
-
-const imageCoverText = css`
-  width: 100%;
-  color: ${theme.colors.gray200};
-  text-shadow: 0px 0px 4px ${theme.colors.shadow60};
-  font-size: 0.8rem;
-  position: absolute;
-  bottom: 0;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  text-align: center;
-  z-index: 0;
-`;
-
-const iconBox = css`
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-`;
-
 const inputContainer = css`
   border-top: 1px solid ${theme.colors.gray300};
   padding-bottom: 1.25rem;
@@ -128,10 +87,6 @@ const styles = {
   inputContainer,
   inputWrapper,
   imageWrapper,
-  imageBox,
-  imageInput,
-  imageCoverText,
-  iconBox,
   input,
 };
